.specify/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain-google-genai==1.0.3
qdrant-client==1.8.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pytest==7.4.3
httpx==0.25.2
pydantic==2.5.0
openai==1.3.4
langchain==0.1.16
cohere==4.2.0
# Note: The 'agents' import in agent.py may require a custom agents library
# that provides Agent, Runner, function_tool, RunConfig, OpenAIChatCompletionsModel
